+++ Atlas.lua
@@ -6,22 +6,22 @@ mounts:RegisterEvent("PLAYER_ENTERING_WORLD")
 function mounts:PLAYER_ENTERING_WORLD()
 	fprint("PLAYER_ENTERING_WORLD")
 	-- JOURNAL OPEN
-	-- if not IsAddOnLoaded("Blizzard_Collections") then
-	-- 	LoadAddOn("Blizzard_Collections")
-	-- end
-	-- CollectionsJournal:Show()
-	-- journal.navBarBtn:Click()
+	if not IsAddOnLoaded("Blizzard_Collections") then
+		LoadAddOn("Blizzard_Collections")
+	end
+	CollectionsJournal:Show()
+	journal.navBarBtn:Click()
 
 
 	-- CONFIG OPEN
 	-- local classConfig = MountsJournalConfig
-	local classConfig = MountsJournalConfigClasses
-	if InterfaceOptionsFrameAddOns:IsVisible() and classConfig:IsVisible() then
-		InterfaceOptionsFrame:Hide()
-	else
-		InterfaceOptionsFrame_OpenToCategory(classConfig.name)
-		if not InterfaceOptionsFrameAddOns:IsVisible() then
-			InterfaceOptionsFrame_OpenToCategory(classConfig.name)
-		end
-	end
+	-- local classConfig = MountsJournalConfigClasses
+	-- if InterfaceOptionsFrameAddOns:IsVisible() and classConfig:IsVisible() then
+	-- 	InterfaceOptionsFrame:Hide()
+	-- else
+	-- 	InterfaceOptionsFrame_OpenToCategory(classConfig.name)
+	-- 	if not InterfaceOptionsFrameAddOns:IsVisible() then
+	-- 		InterfaceOptionsFrame_OpenToCategory(classConfig.name)
+	-- 	end
+	-- end
 end